remove support for really old TS versions
