+++ packages/jimp/package.json
@@ -6,13 +6,6 @@
   "module": "es/index.js",
   "types": "types/index.d.ts",
   "browser": "browser/lib/jimp.js",
-  "typesVersions": {
-    ">=3.1.0-0": {
-      "*": [
-        "types/ts3.1/index.d.ts"
-      ]
-    }
-  },
   "tonicExampleFilename": "example.js",
   "files": [
     "browser",
