+++ packages/jimp/types/test.ts
@@ -3,10 +3,15 @@ import * as Jimp from "jimp";
 const jimpInst: Jimp = new Jimp("test");
 
 // Main Jimp export should already have all of these already applied
+// $ExpectError
 jimpInst.read("Test");
 jimpInst.displace(jimpInst, 2);
 jimpInst.resize(40, 40);
-// $ExpectType 0
+jimpInst.displace(jimpInst, 2);
+jimpInst.shadow((err, val, coords) => {});
+jimpInst.fishEye({ r: 12 });
+jimpInst.circle({ radius: 12, x: 12, y: 12 });
+// $ExpectError
 jimpInst.PNG_FILTER_NONE;
 
 // $ExpectError
@@ -17,9 +22,7 @@ jimpInst.func();
 
 // Main Jimp export should already have all of these already applied
 Jimp.read("Test");
-Jimp.displace(Jimp, 2);
-Jimp.shadow((err, val, coords) => {});
-Jimp.resize(40, 40);
+
 // $ExpectType 0
 Jimp.PNG_FILTER_NONE;
 
@@ -33,12 +36,12 @@ test("can clone properly", async () => {
   const baseImage = await Jimp.read("filename");
   const cloneBaseImage = baseImage.clone();
 
-  // $ExpectType -1
-  cloneBaseImage.PNG_FILTER_AUTO;
+  // $ExpectType number
+  cloneBaseImage._deflateLevel;
 
   test("can handle `this` returns on the core type properly", () => {
-    // $ExpectType -1
-    cloneBaseImage.diff(jimpInst, jimpInst).image.PNG_FILTER_AUTO;
+    // $ExpectType number
+    cloneBaseImage.posterize(3)._quality;
   });
 
   test("can handle `this` returns properly", () => {
@@ -49,16 +52,17 @@ test("can clone properly", async () => {
       .print("a" as any, 2, 2, "a" as any)
       .resize(1, 1)
       .quality(1)
-      .deflateLevel(2).PNG_FILTER_AUTO;
+      .deflateLevel(2)._filterType;
   });
 
   test("can handle imageCallbacks `this` properly", () => {
     cloneBaseImage.rgba(false, (_, jimpCBIn) => {
+      // $ExpectError
       jimpCBIn.read("Test");
       jimpCBIn.displace(jimpInst, 2);
       jimpCBIn.resize(40, 40);
-      // $ExpectType 0
-      jimpCBIn.PNG_FILTER_NONE;
+      // $ExpectType number
+      jimpCBIn._filterType;
 
       // $ExpectError
       jimpCBIn.test;
@@ -73,11 +77,12 @@ test("Can handle callback with constructor", () => {
   const myBmpBuffer: Buffer = {} as any;
 
   Jimp.read(myBmpBuffer, (err, cbJimpInst) => {
+    // $ExpectError
     cbJimpInst.read("Test");
     cbJimpInst.displace(jimpInst, 2);
     cbJimpInst.resize(40, 40);
-    // $ExpectType 0
-    cbJimpInst.PNG_FILTER_NONE;
+    // $ExpectType number
+    cbJimpInst._filterType;
 
     // $ExpectError
     cbJimpInst.test;
@@ -85,17 +90,4 @@ test("Can handle callback with constructor", () => {
     // $ExpectError
     cbJimpInst.func();
   });
-});
-
-test("Can handle appendConstructorOption", () => {
-  Jimp.appendConstructorOption(
-    "Name of Option",
-    (args) => args.hasSomeCustomThing,
-    function (resolve, reject, args) {
-      // $ExpectError
-      this.bitmap = 3;
-      Jimp.resize(2, 2);
-      resolve();
-    }
-  );
 });
