--- packages/jimp/types/ts3.1/tsconfig.json
@@ -1,18 +1,0 @@
-{
-  "compilerOptions": {
-    "module": "commonjs",
-    "lib": ["es6"],
-    "noImplicitAny": true,
-    "noImplicitThis": true,
-    "strictNullChecks": true,
-    "strictFunctionTypes": true,
-    "noEmit": true,
-
-    // If the library is an external module (uses `export`), this allows your test file to import "mylib" instead of "./index".
-    // If the library is global (cannot be imported via `import` or `require`), leave this out.
-    "baseUrl": "../test",
-    "paths": {
-      "mylib": ["."]
-    }
-  }
-}
